Implemented solutions for version2
