+++ workspace-version2/CustomerSrv/src/main/java/webshop/customers/api/BaseResponse.java
@@ -2,9 +2,9 @@ package webshop.customers.api;
 
 import javax.validation.constraints.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import org.hibernate.validator.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BaseResponse {
     @NotNull
