+++ workspace-version2/CustomerSrv/src/main/java/webshop/customers/api/CreditRatingCheckResponse.java
@@ -2,6 +2,7 @@ package webshop.customers.api;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class CreditRatingCheckResponse {
