+++ workspace-version2/OrderProcessSrv/src/main/java/webshop/orderprocess/api/BaseResponse.java
@@ -2,9 +2,9 @@ package webshop.orderprocess.api;
 
 import javax.validation.constraints.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import org.hibernate.validator.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BaseResponse {
     @NotNull
