+++ workspace-version2/OrderProcessSrv/src/main/java/webshop/orderprocess/api/ProductAvailabilityCheckResponse.java
@@ -1,6 +1,7 @@
 package webshop.orderprocess.api;
 
 import javax.validation.constraints.Min;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class ProductAvailabilityCheckResponse {
