+++ workspace-version2/ProductSrv/src/main/java/webshop/products/ServiceApplication.java
@@ -13,13 +13,9 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import io.dropwizard.Application;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
-import webshop.products.db.ProductCategoryRepository;
 import webshop.products.db.ProductRepository;
-import webshop.products.db.WarehouseRepository;
 import webshop.products.health.StandardHealthCheck;
-import webshop.products.resources.ProductCategoryResource;
 import webshop.products.resources.ProductResource;
-import webshop.products.resources.WarehouseResource;
 
 public class ServiceApplication extends Application<ServiceConfiguration> {
 
@@ -38,7 +34,7 @@ public class ServiceApplication extends Application<ServiceConfiguration> {
 
 	@Override
 	public void run(final ServiceConfiguration configuration, final Environment environment) {
-		
+
 		final StandardHealthCheck healthCheck = new StandardHealthCheck(configuration.getDefaultCategoryId());
 
 		// JSON pretty print
@@ -53,22 +49,11 @@ public class ServiceApplication extends Application<ServiceConfiguration> {
 		cors.addMappingForUrlPatterns(EnumSet.allOf(DispatcherType.class), true, "/*");
 
 		environment.healthChecks().register("template", healthCheck);
-		
+
 		// Product resource
 		final ProductRepository productRepository = new ProductRepository();
 		final ProductResource productResource = new ProductResource(configuration.getDefaultCategoryId(),
 				productRepository);
 		environment.jersey().register(productResource);
-		
-		// Category resource
-		final ProductCategoryRepository categoryRepository = new ProductCategoryRepository();
-		final ProductCategoryResource categoryResource = new ProductCategoryResource(categoryRepository);
-		environment.jersey().register(categoryResource);
-		
-		// Warehouse resource
-		final WarehouseRepository warehouseRepository = new WarehouseRepository();
-		final WarehouseResource warehouseResource = new WarehouseResource(warehouseRepository);
-		environment.jersey().register(warehouseResource);
 	}
-
 }
