--- workspace-version2/ProductSrv/src/main/java/webshop/products/resources/WarehouseResource.java
@@ -1,69 +1,0 @@
-package webshop.products.resources;
-
-import javax.ws.rs.DefaultValue;
-import javax.ws.rs.GET;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response.Status;
-
-import com.codahale.metrics.annotation.Timed;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import io.dropwizard.jersey.params.IntParam;
-import io.dropwizard.jersey.params.LongParam;
-import webshop.products.api.BaseResponse;
-import webshop.products.api.ProductAvailabilityCheckResponse;
-import webshop.products.db.WarehouseRepository;
-
-@Path("/products")
-@Produces(MediaType.APPLICATION_JSON)
-public class WarehouseResource {
-	private WarehouseRepository warehouseRepository;
-	private Logger log;
-
-	public WarehouseResource(WarehouseRepository repository) {
-		this.warehouseRepository = repository;
-		this.log = LoggerFactory.getLogger(WarehouseResource.class);
-		log.info("WarehouseResource instantiated...");
-	}
-
-	@Path("/{id}/availability")
-	@GET
-	@Timed
-	public ProductAvailabilityCheckResponse checkProductAvailability(@PathParam("id") LongParam productId,
-			@QueryParam("amount") @DefaultValue("1") IntParam requestedAmount) {
-
-		log.info("Checking availability for product with ID " + productId.get() + " for the amount of "
-				+ requestedAmount.get() + "...");
-		final int availableAmount = warehouseRepository.getAvailableProductAmount(productId.get());
-
-		if (availableAmount == -1) {
-			final String msg = String.format("Product with ID %d does not exist...", productId.get());
-			throw new WebApplicationException(msg, Status.NOT_FOUND);
-		}
-
-		// MAGICALLY RESERVE REQUESTED AMOUNT
-
-		return new ProductAvailabilityCheckResponse(productId.get(), availableAmount, requestedAmount.get());
-	}
-
-	@Path("/{id}/availability")
-	@PUT
-	@Timed
-	public BaseResponse updateProductAvailability(@PathParam("id") LongParam productId,
-			@QueryParam("amount") @DefaultValue("1") IntParam amount) {
-
-		log.info("Setting available amount for product with ID " + productId.get() + " to " + amount.get() + "...");
-		warehouseRepository.setAvailableProductAmount(productId.get(), amount.get());
-
-		return new BaseResponse("OK", 200,
-				"Available amount for product with ID " + productId.get() + " successfully updated.");
-	}
-}
