+++ workspace-version2/ProductSrvFacade/src/main/java/webshop/products/api/BaseResponse.java
@@ -2,9 +2,9 @@ package webshop.products.api;
 
 import javax.validation.constraints.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import org.hibernate.validator.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BaseResponse {
     @NotNull
