+++ workspace-version2/ProductSrvFacade/src/main/java/webshop/products/api/Product.java
@@ -2,6 +2,7 @@ package webshop.products.api;
 
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Min;
+
 import org.hibernate.validator.constraints.NotEmpty;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
