+++ workspace-version2/ProductSrvFacade/src/main/java/webshop/products/api/ProductAvailabilityCheckResponse.java
@@ -1,6 +1,7 @@
 package webshop.products.api;
 
 import javax.validation.constraints.Min;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class ProductAvailabilityCheckResponse {
