+++ workspace-version2/ProductSrvFacade/src/main/java/webshop/products/resources/ProductFacadeResource.java
@@ -18,6 +18,10 @@ import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.Invocation;
 import javax.ws.rs.core.GenericType;
 import javax.ws.rs.core.MediaType;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.codahale.metrics.annotation.Timed;
 
 import io.dropwizard.jersey.params.IntParam;
@@ -26,9 +30,6 @@ import webshop.products.api.BaseResponse;
 import webshop.products.api.Product;
 import webshop.products.api.ProductAvailabilityCheckResponse;
 import webshop.products.api.ProductCategory;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 @Path("/")
 @Produces(MediaType.APPLICATION_JSON)
@@ -97,10 +98,8 @@ public class ProductFacadeResource {
 	@Timed
 	public ProductAvailabilityCheckResponse checkProductAvailability(@PathParam("id") LongParam productId,
 			@QueryParam("amount") @DefaultValue("1") IntParam requestedAmount) {
-		
-		// TODO Ex2, Task3: Change to use the WAREHOUSE_SRV_ENDPOINT variable
 		final Invocation.Builder request = restClient
-				.target(PRODUCT_SRV_ENDPOINT + "/products/" + productId + "/availability?amount=" + requestedAmount)
+				.target(WAREHOUSE_SRV_ENDPOINT + "/products/" + productId + "/availability?amount=" + requestedAmount)
 				.request();
 		return request.get(ProductAvailabilityCheckResponse.class);
 	}
@@ -110,10 +109,8 @@ public class ProductFacadeResource {
 	@Timed
 	public BaseResponse updateProductAvailability(@PathParam("id") LongParam productId,
 			@QueryParam("amount") @DefaultValue("1") IntParam amount) {
-		
-		// TODO Ex2, Task3: Change to use the WAREHOUSE_SRV_ENDPOINT variable
 		final Invocation.Builder request = restClient
-				.target(PRODUCT_SRV_ENDPOINT + "/products/" + productId + "/availability?amount=" + amount).request();
+				.target(WAREHOUSE_SRV_ENDPOINT + "/products/" + productId + "/availability?amount=" + amount).request();
 		return request.put(Entity.json(""), BaseResponse.class);
 	}
 
@@ -123,8 +120,7 @@ public class ProductFacadeResource {
 	@GET
 	@Timed
 	public List<ProductCategory> getCategories(@QueryParam("limit") @DefaultValue("20") IntParam limit) {
-		// TODO Ex2, Task3: Change to use the CATEGORY_SRV_ENDPOINT variable
-		final Invocation.Builder request = restClient.target(PRODUCT_SRV_ENDPOINT + "/categories?limit=" + limit)
+		final Invocation.Builder request = restClient.target(CATEGORY_SRV_ENDPOINT + "/categories?limit=" + limit)
 				.request();
 		return request.get(new GenericType<List<ProductCategory>>() {
 		});
@@ -134,8 +130,7 @@ public class ProductFacadeResource {
 	@GET
 	@Timed
 	public ProductCategory getCategoryById(@PathParam("id") LongParam categoryId) {
-		// TODO Ex2, Task3: Change to use the CATEGORY_SRV_ENDPOINT variable
-		final Invocation.Builder request = restClient.target(PRODUCT_SRV_ENDPOINT + "/categories/" + categoryId)
+		final Invocation.Builder request = restClient.target(CATEGORY_SRV_ENDPOINT + "/categories/" + categoryId)
 				.request();
 		return request.get(ProductCategory.class);
 	}
@@ -144,8 +139,7 @@ public class ProductFacadeResource {
 	@POST
 	@Timed
 	public BaseResponse createCategory(@NotNull @Valid ProductCategory category) {
-		// TODO Ex2, Task3: Change to use the CATEGORY_SRV_ENDPOINT variable
-		final Invocation.Builder request = restClient.target(PRODUCT_SRV_ENDPOINT + "/categories").request();
+		final Invocation.Builder request = restClient.target(CATEGORY_SRV_ENDPOINT + "/categories").request();
 		return request.post(Entity.json(category), BaseResponse.class);
 	}
 
@@ -154,8 +148,7 @@ public class ProductFacadeResource {
 	@Timed
 	public BaseResponse updateCategory(@PathParam("id") LongParam categoryId,
 			@NotNull @Valid ProductCategory category) {
-		// TODO Ex2, Task3: Change to use the CATEGORY_SRV_ENDPOINT variable
-		final Invocation.Builder request = restClient.target(PRODUCT_SRV_ENDPOINT + "/categories/" + categoryId)
+		final Invocation.Builder request = restClient.target(CATEGORY_SRV_ENDPOINT + "/categories/" + categoryId)
 				.request();
 		return request.put(Entity.json(category), BaseResponse.class);
 	}
@@ -164,8 +157,7 @@ public class ProductFacadeResource {
 	@DELETE
 	@Timed
 	public BaseResponse deleteCategory(@PathParam("id") LongParam categoryId) {
-		// TODO Ex2, Task3: Change to use the CATEGORY_SRV_ENDPOINT variable
-		final Invocation.Builder request = restClient.target(PRODUCT_SRV_ENDPOINT + "/categories/" + categoryId)
+		final Invocation.Builder request = restClient.target(CATEGORY_SRV_ENDPOINT + "/categories/" + categoryId)
 				.request();
 		return request.delete(BaseResponse.class);
 	}
