+++ workspace-version2/WarehouseSrv/src/main/java/webshop/warehouse/api/Product.java
@@ -2,6 +2,7 @@ package webshop.warehouse.api;
 
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Min;
+
 import org.hibernate.validator.constraints.NotEmpty;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
