+++ workspace-version2/WarehouseSrv/src/main/java/webshop/warehouse/db/WarehouseRepository.java
@@ -1,11 +1,35 @@
 package webshop.warehouse.db;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class WarehouseRepository {
 
-	// TODO Ex2, Task2: Replace this class with warehouse (product availability) repository class from the ProductSrv
-	
-	public WarehouseRepository() {
+	private Map<Long, Integer> warehouse;
 
+	public WarehouseRepository() {
+		this.warehouse = new HashMap<Long, Integer>() {
+			private static final long serialVersionUID = 1L;
+			{
+				put((long) 1, 5);
+				put((long) 2, 4);
+				put((long) 3, 3);
+				put((long) 4, 2);
+				put((long) 5, 1);
+			}
+		};
 	}
 
+	public int getAvailableProductAmount(long productId) {
+		int availableAmount = -1;
+		if (warehouse.get(productId) != null) {
+			availableAmount = warehouse.get(productId);
+		}
+
+		return availableAmount;
+	}
+
+	public void setAvailableProductAmount(long productId, int amount) {
+		warehouse.put(productId, amount);
+	}
 }
