+++ workspace-version2/WarehouseSrv/src/main/java/webshop/warehouse/messaging/KafkaListener.java
@@ -37,6 +37,7 @@ public class KafkaListener implements Runnable {
 
 	@Override
 	public void run() {
+		@SuppressWarnings("resource")
 		final Consumer<String, Product> consumer = new KafkaConsumer<String, Product>(kafkaProps);
 		consumer.subscribe(Arrays.asList(KAFKA_TOPIC_NAME));
 		while (true) {
@@ -44,9 +45,10 @@ public class KafkaListener implements Runnable {
 			for (ConsumerRecord<String, Product> message : messages) {
 				Product createdProduct = message.value();
 
-				// TODO Ex3, Task3: Increase the available amount of the new product to 10 using 'this.warehouseResource.updateProductAvailability()'
-				
-				
+				// Stock up on 10 copies of the new product
+				LongParam productId = new LongParam(String.valueOf(createdProduct.getId()));
+				IntParam amount = new IntParam("10");
+				this.warehouseResource.updateProductAvailability(productId, amount);
 			}
 		}
 	}
